+++ previous-versions/assess/src/sdk/models/operations/generateloantransactions.ts
@@ -7,15 +7,15 @@ import * as shared from "../shared";
 import { AxiosResponse } from "axios";
 
 /**
- * Data source type.
+ * Data source type
  */
-export enum GetLoanSummarySourceType {
+export enum GenerateLoanTransactionsSourceType {
     Banking = "banking",
     Commerce = "commerce",
     Accounting = "accounting",
 }
 
-export class GetLoanSummaryRequest extends SpeakeasyBase {
+export class GenerateLoanTransactionsRequest extends SpeakeasyBase {
     /**
      * Unique identifier for a company.
      */
@@ -23,13 +23,13 @@ export class GetLoanSummaryRequest extends SpeakeasyBase {
     companyId: string;
 
     /**
-     * Data source type.
+     * Data source type
      */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=sourceType" })
-    sourceType: GetLoanSummarySourceType;
+    sourceType: GenerateLoanTransactionsSourceType;
 }
 
-export class GetLoanSummaryResponse extends SpeakeasyBase {
+export class GenerateLoanTransactionsResponse extends SpeakeasyBase {
     /**
      * HTTP response content type for this operation
      */
@@ -41,12 +41,6 @@ export class GetLoanSummaryResponse extends SpeakeasyBase {
      */
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
-
-    /**
-     * OK
-     */
-    @SpeakeasyMetadata()
-    loanSummary?: shared.LoanSummary;
 
     /**
      * HTTP response status code for this operation