+++ previous-versions/assess/src/sdk/sdk.ts
@@ -55,7 +55,7 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "1.0";
-    sdkVersion = "0.35.1";
+    sdkVersion = "0.35.2";
     genVersion = "2.131.1";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {